requirements.txt
requests == 2.32.3
praw == 7.7.1
tqdm == 4.66.5
pyspark == 3.5.3
pandas == 2.0.3
PyYAML == 6.0.1
kafka-python == 2.0.2
apache-airflow == 2.8.4
python-dotenv == 1.0.1
google-cloud-storage == 2.18.2
google-auth == 2.35.0
pyarrow == 17.0.0
# pip install pyspark-stubs --no-deps